pyproject.toml
[tool.poetry]
name = "django-homework"
version = "0.1.0"
description = "django-homework hillel"
authors = ["netscro <kovalev_evgeniy@list.ru>"]

[tool.poetry.dependencies]
python = "^3.7.3"
Django = "3.1.4"
gender-guesser = "0.4.0"
celery = "5.0.5"
django-filter = "^2.4.0"
SQLAlchemy = "1.3.22"
requests = "2.25.1"
django-redis = "4.12.1"
gunicorn = "20.0.4"
Pillow = "^8.1.1"
django-extensions = "^3.1.1"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.2"
freezegun = "^1.1.0"

[tool.poetry.dev-dependencies]
flake8 = "3.8.4"
Faker = "5.0.2"
flake8-builtins = "1.5.3"
flake8-isort = "4.0.0"
flake8-print = "4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
